refactors appearancecomp
